TouchButton::TouchButton(custom) pass: restore spacing ABI (84.3->100 x2, linked 2679->2682)
Retyping the authentic signed-short getter also makes ImageFontGetSpacing byte/linked exact (72.7->100), accounting for the third linked gain.

+++ src/engine/core/AbyssEngine.cpp
@@ -1570,7 +1570,7 @@ namespace AbyssEngine {
 }
 
 namespace AbyssEngine {
-    int ImageFontGetSpacing(ImageFont *font) {
+    short ImageFontGetSpacing(ImageFont *font) {
         short v = (font == 0) ? (short) 0 : font->spacing;
         return (int) v;
     }
